v4l2: add VIDTYPE_COMPRESS in vf for dw mode 1/2/4 [1/1]
PD#SWPL-27197

Problem:
When double write mode is set to 1/2/4,  DRM will show the content
inside vframe.  This is because VIDTYPE_COMPRESS is not set. When dw
mode is set to 3, both linear and compressed buffer will be passed to
display, and display will decide which to show according to current
downscaled ratio.

Solution:
On linux v4l2, the requirement is compressed buffer is preferred on
video plane. And downscale is done on linear buffer.

Verify:
U212 + RDK4

Change-Id: Iea93ad9cbb9f5caa397b943acbd2f390378ed08c
Signed-off-by: Song Zhao <song.zhao@amlogic.com>