media_module: h264 and mpeg2 pip playback, h264 show mosaic picture [1/1]
PD#SWPL-40045

Problem:
When the IQIDCT_CONTROL register is not saved
because CSD data is not found at playback start,
next time to decode, the IQIDCT_CONTROL register's
value is wrong, cause this problem.

Solution:
When the IQIDCT_CONTROL register is not saved,
set IQIDCT_CONTROL register the initialized value(0x200).

Verify:
SC2

Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>
Change-Id: I8af14845f953f9e6e95b5f766f4bcec6b513b871