h265: CB1 front and back concurrency causes crash problems [1/1]
PD#SWPL-198847

Problem:
The front is an error frame, which the back discards without
decoding. When the front is processing its buf, the back marks
its state as complete ahead of time, causing the front frame
to accidentally acquire the unprocessed buf when decoding it
and set it to null, thus triggering a null pointer operation.

Solution:
After back has finished processing the frame, set the buf state to done,

Verify:
S5

Change-Id: I77e5353b591a71c1649edffcd27f79f173bac5f2
Signed-off-by: xing.xu <xing.xu@amlogic.com>