h264: CB1 resolution change & seek concurrently, causes dmc violation [1/1]
PD#SWPL-202956

Problem:
Resolution switching occurs concurrently with the seek action,
causing the switching process to be interrupted. As a result,
the decoder fails to complete the resolution switch properly,
leading to abnormal operation.

Solution:
A fault-tolerance mechanism has been added.
If the resolution switch is interrupted by a seek action,
a reset mechanism is triggered to restore the resolution
to its previous state, preventing decoder malfunctions.

Verify:
T5M

Change-Id: Iaea791d35a52a33457055c368a46b99d7953b05d
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>