di-backend: CB2 I/P stream switching crash issues. [1/1]
PD#SWPL-202195
PD#SWPL-202194

Problem:
1.During the time from when the decoder processes
the new resolution header to when it reports the resolution
to the upper layer,if the upper layer queues the frame and the decoder
I/P mode has not been switched,the frame is inserted into the ubuf list.
Subsequent access to the ubuf list may cause problems
because the frame may be released. If the decoder I/P mode is switched,
a null pointer may appear when the frame queues in.
2.The DI interface crashes when I/P stream are switched.
This is because the DI buf mgr context is not cleaned up
when I/P resolution is switched.

Solution:
1.When I/P streams are switched, the ubuf list is cleared at streamoff
and if a frame before the queue changes resolution is received at this time,
the frame is ignored.
2.When I/P streams are switched, di buf mgr will be destroyed and
then created to ensure that di buf mgr is clean after switching.

Verify:
ohm

Change-Id: Iebd53f444d2e4f55482e6a8bc708a7ca0ccf3b00
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>