mpeg2: CB2 fix stuck issue when 1-package-1-field and field is dropped [1/1]
PD#SWPL-193323

Problem:
When one field is dropped, the current field will be paired with the
field of next picture, and there maybe no failure or error, but show mosaic.
It will continue to pair incorrectly, and won't fail until the P/B
field is paired with I field, then dropping frame all the time and stuck.

Solution:
In the ucode, when the case of unpaired fields is checked,
mark it unpaired and error.
In the driver, not dirty the 2nd field package data, decode it again,
and try to pair with the next field.

Verify:
ohm

SourceCode:
ucode:
0.4.181-g8bf8ee6
I762bf
Id7df5
I14824
I644d4
Iaca58

Change-Id: I544d20598bfb829e1c41792300e1bc0813a931b5
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>