vdec: CF2 Add AV1 error-frame dropping in fence mode [1/1]
PD#SWPL-203470

Problem:
in fence mode, the frame buffer is posted to display when
frame header is decoded. But only the frame body is decoded, we can know
the frame is error or not.
1)the error frame can not be dropped and it will show mosaic.
2)the buffer was posted, but the fence is not trigger if error-frame.

Solution:
a) complete AV1 error-stream detection on fb_non-v4l and t5d.
b) in fence mode, drop error frame in display.
Display will keep the showing frame and stop keeping when next
showenable frame coming. It needs decoder to check the fence statuts and
refcount to get the buffer status.
1)decoder mark error frame before triggerring fence signal.
2)when the buffer is posted to the backend, but fence is not signaled,
triggering fence signal once in work.

Verify:
SC2

Change-Id: Icc2d51fe8ebc73126e1895b9cfdfae00f11e66c2
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>