h264: CB2 fix timeout problems in mmu mode. [1/1]
PD#SWPL-212621

Problem:
When mmu decoding is turned on in stream mode, it will retry to decode when
there is insufficient data. This will cause the decoding driver to wait for
some hevc register status timeout (200ms) and cause playback problems.

Solution:
Since the reset is added to wait for the hevc module to be idle before
decoding again, the h264 decoding driver does not need to check the status of
the hevc register when an exception occurs.

Verify:
GXLX4

Change-Id: Iaf006811b85477c26d8c73170592b99f25fcc11e
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>