v4ldec: CB0 Too many allocations for YUV buffer of interlace streams. [1/1]
PD#SWPL-210313

Problem:
Over-allocation of YUV buffers leads to insufficient memory and crash.
The reason for over-allocation of YUV buffers is that the decoder
first applies for YUV buffers and then pairs them,
but the original uvm dma buffer has not been recycled,
resulting in over-allocation of buffers.

Solution:
Limit the number of YUV DMA buffer requests to
no more than the size of DPB size.

Verify:

Change-Id: I3eb645a733663d340441fa87f042e962f268713a
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>