v4l: CF2 Playback stuttering occurs with specific video sources [1/1]
PD#SWPL-214359

Problem:
In the new CSD, there is a POC discontinuity in this I-frame.
After the decoder finishes decoding and outputting this frame,
the DPB becomes full. As a result, it triggers a DPB flush,
causing the P/B frames of this GOP to fail to decode properly,
leading to playback stutter.

Solution:
The set_params for this frame includes a DPB flush operation.

Verify:
s6

Change-Id: Ifc0ff9c000f5e84e05f86aa5f0b691a16d727e83
Signed-off-by: xing.xu <xing.xu@amlogic.com>