vdec: CB2 fix buffer leak in fence mode when error frame early put back [1/1]
PD#SWPL-213121

Problem:
1. In fence mode, the posted-buffer is early put back by omx,
so the vf will be stored and buffer will not be recycled
until the fence is triggered.
When the frame is error, the fence status will be marked as -1
before triggering.
Finally the fence with error status is not recycled and buffer is not
recycled, causing buffer leak.
2. config_decode_buf will goto reset if error frame and
not posting error-frame, in fence mode, it will be black screen
or stuck for a long time, because less I frames are in low latency stream.

Solution:
1.In fence mode, recycle the fence with error status and buffer.
2.In fence mode, do not goto reset when not posting error-frame.

Verify:
T950S

Change-Id: I7887bb70b78834fc75aeccbd100518c0e475c520
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>