cc: CF2 High latency in obtaining main channel CC data for subtitles [1/1]
PD#SWPL-215968

Problem:
In amstream, the FCC (Fast Channel Change) path blocks the main channel.
The video ID passed from the upper layer is -1, which coincides with
the default value of userdata->id[i], also initialized as -1.
This causes an unintended match.

Solution:
If the video ID received from the upper layer is -1,
it should be treated as invalid and ignored to
prevent incorrect channel matching.

Verify:
sc2

Change-Id: I91977ec1ecbb375512b4eeec5fed7718908bb234
Signed-off-by: xing.xu <xing.xu@amlogic.com>