v4ldec: CB0 Memory migration fails due to memory page pollution. [1/1]
PD#SWPL-215309

Problem:
The fake buffer is not immediately replaced with a new buffer
of actual size after release,
resulting in the buffer being unmapped repeatedly after exit.

Solution:
Release the fake buffer only when you are sure that
you can replace it with a buffer of the actual size.

Verify:
t6d

Change-Id: I2829c9003bc5d7cf9588ec6dfe413521ec19b81a
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>