v4ldec: CB0 Kernel crash when accessing workqueue. [1/1]
PD#SWPL-221916

Problem:
The workqueue failed to be created when the broadcast started,
and when it was exited,
it flushed the queue, causing a null pointer to crash.

Solution:
When a v4ldec instance exits, it checks if the workqueue exists.

Verify:
ohm

Change-Id: Ia24f6c1adede8d2559a0e6d97a96be6e8a63d000
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>
(cherry picked from commit b6a949de72bbfb214b1f4d28d5f0f85185ba8060)