h264: CB2 non-v4l playback is stuck [1/1]
PD#SWPL-219185

Problem:
After the stream loops back, the buffer is full.
As a result, get free buf fails, causing the issue.

Solution:
When get free buf fails, the decoder immediately
terminates the decoding of the current frame.
During the run process, check if buf_alloc_fail
is set to 1, and if so, perform a buffer reset operation.

Verify:
GXLX4

Change-Id: I1203220a68f5e4c3f4e7877cc86710ffed6423e5
Signed-off-by: xing.xu <xing.xu@amlogic.com>