avs/mpeg4: CB0 avs error stream crash. [1/1]
PD#SWPL-219099

Problem:
When avs processes some error frames,
it recycles the buffer and sets it to a free available state,
causing two frames to use one buffer at the same time,
resulting in a rotation problem.

Solution:
When processing error frames, the buffer is recycled
and must be in a bound state.

Verify:
Ross

Change-Id: I737d7f46ce5cbef80aaf896d0dc19599cd69db32
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>