v4ldec: CB2 Interlaced stream seeking stuck. [1/1]
PD#SWPL-220653
PD#SWPL-219116

Problem:
UVM DMABUF pairing is only performed
when the buffer is sent to the decoder.
Too many unpaired UVM DMABUFs will cause the decoder
to have no combined UVM DMABUF available, resulting in a stuck.

Solution:
When the decoder run_ready is on, it detects whether
there is a qualified UVM DMABUF that can be paired.
If so, it pairs it so that the UVM DMABUF can be combined as soon as possible.

Verify:
Ross

Change-Id: Ide4f4a7dcb7aabb993f5c653a3e3dff766a4c98a
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>