h264: CB2 interlace playback stuck [1/1]
PD#SWPL-220274
PD#OTT-80588

Problem:
Occasionally, after the top field is output,
the frame is output again during the combine
process. This causes the vf ref to be decremented
to -1, leading to a stall.

Solution:
The combine function uses the last_pic pointer
to store the previous field's picture. After
the top field is output, set last_pic to NULL.

Verify:
ohm

Change-Id: Iaa31bab93b6c38225f2d9841125c25eb139fd8f1
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>