v4ldec: CB0 Sei buffer crash issue. [1/1]
PD#SWPL-233533

Problem:
For vdec core interlaced stream, the sei buffer follows the aml buffer,
but the aml buffer will be released when the frame is output,
resulting in the sei buffer being unable to be recycle and a crash.

Solution:
After the capture buffer returns to the decoder,
call back the sei buffer recycle interface to release the aml buffer.

Verify:
ohm

Change-Id: I7c8d23ebcac3e9003acc035f4ae0803c02e09f31
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>