avbcd: CB2 Some screenshots of the streams have green stripes. [1/1]
PD#SWPL-242134

Problem:
When configuring the ge2d parameters,
the src para parameter should be configured with
the actual size of the source YUV region.
In reality, the alignment size is configured,
but this will cause some resolutions to have extra green stripes
in the screenshot.

Solution:
The `src para` parameter is configured according to
the actual YUV dimensions (the dimensions after cropping, not aligned).

Verify:
S6

Change-Id: I3f5b196bab53529c85587d23f85e1fe2532f90e3
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>